+++ pkg/collector/parser/receiver_generic_test.go
@@ -77,6 +77,9 @@ func TestDownstreamParsers(t *testing.T) {
 		{"zipkin-scribe", "zipkin-scribe", "__zipkinscribe", 9410, parser.NewZipkinScribeReceiverParser},
 		{"fluentforward", "fluentforward", "__fluentforward", 8006, parser.NewFluentForwardReceiverParser},
 		{"statsd", "statsd", "__statsd", 8125, parser.NewStatsdReceiverParser},
+		{"influxdb", "influxdb", "__influxdb", 8086, parser.NewInfluxdbReceiverParser},
+		{"splunk-hec", "splunk-hec", "__splunk_hec", 8088, parser.NewSplunkHecReceiverParser},
+		{"awsxray", "awsxray", "__awsxray", 2000, parser.NewAWSXrayReceiverParser},
 	} {
 		t.Run(tt.receiverName, func(t *testing.T) {
 			t.Run("builds successfully", func(t *testing.T) {