+++ pkg/collector/reconcile/deployment_test.go
@@ -24,7 +24,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
-	"github.com/open-telemetry/opentelemetry-operator/api/v1alpha1"
+	"github.com/open-telemetry/opentelemetry-operator/api/collector/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/targetallocator"
 )
@@ -121,7 +121,7 @@ func TestExpectedDeployments(t *testing.T) {
 		createObjectIfNotExists(t, "test-targetallocator", &expectedTADeploy)
 		orgUID := expectedTADeploy.OwnerReferences[0].UID
 
-		updatedParam, err := newParams("test/test-img")
+		updatedParam, err := newParams("test/test-img", "")
 		assert.NoError(t, err)
 		updatedDeploy := targetallocator.Deployment(updatedParam.Config, logger, param.Instance)
 		*updatedDeploy.Spec.Replicas = int32(3)
@@ -144,7 +144,7 @@ func TestExpectedDeployments(t *testing.T) {
 		createObjectIfNotExists(t, "test-targetallocator", &expectedTADeploy)
 		orgUID := expectedTADeploy.OwnerReferences[0].UID
 
-		updatedParam, err := newParams("test/test-img")
+		updatedParam, err := newParams("test/test-img", "")
 		assert.NoError(t, err)
 		updatedDeploy := targetallocator.Deployment(updatedParam.Config, logger, updatedParam.Instance)